+++ inst/templates/main_script.R
@@ -125,19 +125,19 @@ components <- generateComponents(fGroups, "nontarget", ionization = "{{ polarity
 {{ endCodeBlock() }}
 {{ optionalCodeBlock("CSV" %in% reportFormats) }}
 reportCSV(fGroups, path = "report", reportFeatures = FALSE, formConsensus = {{ if (formulaOpts$algo != "") "formulas" else "NULL" }},
-          compounds = {{ if (identOpts$algo != "") "compounds" else "NULL" }}, compoundNormalizeScores = TRUE,
+          compounds = {{ if (identOpts$algo != "") "compounds" else "NULL" }}, compoundNormalizeScores = "max",
           components = {{ if (identOpts$algo != "") "components" else "NULL" }})
 
 {{ endCodeBlock() }}
 {{ optionalCodeBlock("PDF" %in% reportFormats) }}
 reportPDF(fGroups, path = "report", reportFGroups = TRUE, formConsensus = {{ if (formulaOpts$algo != "") "formulas" else "NULL" }}, reportFormulaSpectra = TRUE,
-          compounds = {{ if (identOpts$algo != "") "compounds" else "NULL" }}, compoundNormalizeScores = TRUE,
+          compounds = {{ if (identOpts$algo != "") "compounds" else "NULL" }}, compoundNormalizeScores = "max",
           components = {{ if (identOpts$algo != "") "components" else "NULL" }}, MSPeakLists = {{ if (formulaOpts$algo != "" || identOpts$algo != "") "plists" else "NULL" }})
 
 {{ endCodeBlock() }}
 {{ optionalCodeBlock("MD" %in% reportFormats) }}
 reportMD(fGroups, path = "report", reportPlots = c("chord", "venn", "upset", "eics", "formulas"), formConsensus = {{ if (formulaOpts$algo != "") "formulas" else "NULL" }},
-         compounds = {{ if (identOpts$algo != "") "compounds" else "NULL" }}, compoundNormalizeScores = TRUE,
+         compounds = {{ if (identOpts$algo != "") "compounds" else "NULL" }}, compoundNormalizeScores = "max",
          components = {{ if (componentOpts$algo != "") "components" else "NULL" }}, MSPeakLists = {{ if (formulaOpts$algo != "" || identOpts$algo != "") "plists" else "NULL" }},
          selfContained = TRUE, openReport = TRUE)